// amplify/backend/types/amplify-dependent-resources-ref.d.ts
export type AmplifyDependentResourcesAttributes = {
  "api": {
    "courseflexibility": {
      "GraphQLAPIEndpointOutput": "string",
      "GraphQLAPIIdOutput": "string",
      "GraphQLAPIKeyOutput": "string"
    }
  },
  "auth": {
    "courseflexibilityf504702c": {
      "AppClientID": "string",
      "AppClientIDWeb": "string",
      "IdentityPoolId": "string",
      "IdentityPoolName": "string",
      "UserPoolArn": "string",
      "UserPoolId": "string",
      "UserPoolName": "string"
    }
  },
  "storage": {
    "courseFlexibilityFileStorage": {
      "BucketName": "string",
      "Region": "string"
    }
  }
}